Modify getConferenceParticipants() to return copy of the conference data.
In ImsConference#updateConferenceAfterCreation, getConferenceParticipants
is called to get a copy of the CEP data.  This, however, happens on a
different thread than the update of that data.

The ImsCall#getConferenceParticipants() method already synchronizes the
return of the data on a lock which would prevent it from being updated
while the "get" method is called, however, once the reference is returned
there is no guarantee the underlying list won't be changed.

Bug: 30861872
Change-Id: I27a1aa35299d36588c73bc47710e129e5e537a6a